Normalize Python runtime path verification
